feat: add monthly payment calculator
